fix: cast chunk to output dtype in scatter_active to avoid dtype mismatch under autocast
The backbone upcasts its output to fp32 while the scale (output) stays in
fp16/bf16 under AMP, making torch.scatter fail. Cast chunk to output.dtype
and add a regression test.

+++ kairos/modeling.py
@@ -395,7 +395,7 @@ class KairosScaleRouter(nn.Module):
         D = output.shape[-1]
         idx = positions.unsqueeze(-1).expand(-1, -1, D)
         current = torch.gather(output, 1, idx)
-        values = torch.where(pad_mask.unsqueeze(-1), chunk, current)
+        values = torch.where(pad_mask.unsqueeze(-1), chunk.to(output.dtype), current)
         return output.scatter(1, idx, values)
